Fixed OS X compilation; in rather ugly manner, using %llx + uulong to replace uint64_t as it seems to be different between Linux and OS X, argh.

+++ src/hcp_i.h
@@ -6,7 +6,7 @@
  * Copyright (c) 2013 cisco Systems, Inc.
  *
  * Created:       Wed Nov 20 13:56:12 2013 mstenber
- * Last modified: Mon Dec  2 14:10:51 2013 mstenber
+ * Last modified: Mon Dec  2 15:10:56 2013 mstenber
  * Edit time:     104 min
  *
  */
@@ -208,9 +208,9 @@ void hcp_self_flush(hcp_node n);
 /* Various hash calculation utilities. */
 void hcp_calculate_hash(const void *buf, int len, hcp_hash dest);
 void hcp_calculate_network_hash(hcp o);
-static inline uint64_t hcp_hash64(hcp_hash h)
+static inline unsigned long long hcp_hash64(hcp_hash h)
 {
-  return *((uint64_t*)h);
+  return *((unsigned long long*)h);
 }
 
 /* Utility functions to send frames. */
